fix some warnings

+++ KS3SDKIOS/KS3YunSDK/KS3YunSDK/KS3YunSDK/KSYS3/KSYModel/Object/KS3FileUploader.m
@@ -88,7 +88,7 @@
     if (_uploadNum == _totalNum) {
         data = [_fileHandle readDataToEndOfFile];
     }else {
-        data = [_fileHandle readDataOfLength:partLength];
+        data = [_fileHandle readDataOfLength:(NSUInteger)partLength];
         [_fileHandle seekToFileOffset:partLength*(_uploadNum)];
     }
